Tercer etapa de refactorizacion
